Homoscedastic Uniform Gaussian synthetic test data.

Homoscedastic Uniform Gaussian synthetic data frame with 7 variables and 5000 measurements for test/evaluation. This dataset was generated to build a simulation scenario to evaluate NeuralGAM.

data(test)

Format

test is a data frame with 7 variables (columns). The first three columns of the data frame contains the covariates X0, X1, X2. The next three columns of the data frame contains the corresponding f(X) for each covariate The last column of the data frame contains the response variable. contains the response variable

Details

We considered the predictor: \(\eta = \alpha + \sum^3_{j=1} f_j(X_j)\), with

$$ f_j(X_j) = \begin{cases} X_j^2 \hspace{1.4cm} \text{ if } j=1 \\ 2X_j \hspace{1.3cm} \text{ if } j=2 \\ \sin{X_j} \hspace{1cm} \text{ if } j=3, \end{cases} $$ \(\alpha = 2\), and covariates \(X_1, X_2, X_3\) drawn from an uniform distribution \(U\left[-2.5, 2.5\right]\) and response variable \(Y = \eta + \varepsilon\) where \(\varepsilon\) is an homoscedastic error term distributed in accordance to a \(N(0,\sigma(x)\)

Author

Ines Ortega-Fernandez, Marta Sestelo.

Examples

library(NeuralGAM)
data(test)
head(test)
#>           X0          X1         X2      f(X0)      f(X1)      f(X2)          y
#> 1  1.2519661 -0.85707121  2.1715824 -0.4767904 -1.6684692 -1.6684692 -1.4413401
#> 2 -2.1458140 -1.00359080 -0.1053674  2.5603083 -1.9615084 -1.9615084  0.4733860
#> 3  0.1686035 -0.08316541 -0.9876430 -2.0157824 -0.1206576 -0.1206576 -2.7854777
#> 4  2.4996863 -0.66870777  0.2114118  4.2042223 -1.2917423 -1.2917423  3.3990273
#> 5  0.2358134  0.10774450 -0.4333229 -1.9886016  0.2611622  0.2611622 -2.2356669
#> 6 -2.4260156 -1.86506489  1.5200782  3.8413420 -3.6844566 -3.6844566  0.2331869

X_train <- test[c("X0", "X1", "X2")]
fs_train <- test[c("f(X0)", "f(X1)", "f(X2)")]
y_train <- test["y"]