Homoscedastic Uniform Gaussian synthetic training data.

Homoscedastic Uniform Gaussian synthetic data frame with 7 variables and 24500 measurements for training. This dataset was generated to build a simulation scenario to evaluate NeuralGAM.

data(train)

Format

train is a data frame with 7 variables (columns). The first three columns of the data frame contains the covariates \(X_1, X_2, X_3\) The next three columns of the data frame contains the corresponding \(f(X)\) for each covariate The last column of the data frame contains the response variable \(y\)

Details

We considered the predictor: \(\eta = \alpha + \sum^3_{j=1} f_j(X_j)\), with

$$ f_j(X_j) = \begin{cases} X_j^2 \hspace{1.4cm} \text{ if } j=1 \\ 2X_j \hspace{1.3cm} \text{ if } j=2 \\ \sin{X_j} \hspace{1cm} \text{ if } j=3, \end{cases} $$ \(\alpha = 2\), and covariates \(X_1, X_2, X_3\) drawn from an uniform distribution \(U\left[-2.5, 2.5\right]\) and response variable \(Y = \eta + \varepsilon\) where \(\varepsilon\) is an homoscedastic error term distributed in accordance to a \(N(0,\sigma(x))\)

Author

Ines Ortega-Fernandez, Marta Sestelo.

Examples

library(NeuralGAM)
data(train)
head(train)
#>           X0         X1         X2        f(X0)     f(X1)     f(X2)          y
#> 1 -0.7351906 -1.7092095  1.6156861 -1.538614934 -3.430178 -3.430178 -5.1207630
#> 2 -0.9346380  1.8562425 -2.0429157 -1.205572066  3.700726  3.700726  1.6531381
#> 3 -1.4816599  0.5976165 -2.3795450  0.116195738  1.183474  1.183474  0.1776291
#> 4  0.5945769  2.2365318 -0.4403814 -1.725598564  4.461305  4.461305  2.4424721
#> 5  1.4386284  1.7939171 -2.3460923 -0.009468417  3.576076  3.576076  2.8480231
#> 6 -0.1694944 -1.0257769  1.8405410 -2.050391860 -2.063313 -2.063313 -2.1984809

X_train <- train[c("X0", "X1", "X2")]
fs_train <- train[c("f(X0)", "f(X1)", "f(X2)")]
y_train <- train$y